% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__AcceleratorConfig}
\alias{GoogleCloudMlV1__AcceleratorConfig}
\title{GoogleCloudMlV1__AcceleratorConfig Object}
\usage{
GoogleCloudMlV1__AcceleratorConfig(type = NULL, count = NULL)
}
\arguments{
\item{type}{The type of accelerator to use}

\item{count}{The number of accelerators to attach to each machine running the job}
}
\value{
GoogleCloudMlV1__AcceleratorConfig object
}
\description{
GoogleCloudMlV1__AcceleratorConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a hardware accelerator request config. Note that the AcceleratorConfig can be used in both Jobs and Versions. Learn more about \href{/ml-engine/docs/using-gpus}{accelerators for training} and \href{/ml-engine/docs/machine-types-online-prediction#gpus}{accelerators for online prediction}.
}
\concept{GoogleCloudMlV1__AcceleratorConfig functions}
