% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__AddTrialMeasurementRequest}
\alias{GoogleCloudMlV1__AddTrialMeasurementRequest}
\title{GoogleCloudMlV1__AddTrialMeasurementRequest Object}
\usage{
GoogleCloudMlV1__AddTrialMeasurementRequest(measurement = NULL)
}
\arguments{
\item{measurement}{Required}
}
\value{
GoogleCloudMlV1__AddTrialMeasurementRequest object
}
\description{
GoogleCloudMlV1__AddTrialMeasurementRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the AddTrialMeasurement service method.
}
\seealso{
Other GoogleCloudMlV1__AddTrialMeasurementRequest functions: 
\code{\link{projects.locations.studies.trials.addMeasurement}()}
}
\concept{GoogleCloudMlV1__AddTrialMeasurementRequest functions}
