% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__AutomatedStoppingConfig}
\alias{GoogleCloudMlV1__AutomatedStoppingConfig}
\title{GoogleCloudMlV1__AutomatedStoppingConfig Object}
\usage{
GoogleCloudMlV1__AutomatedStoppingConfig(
  decayCurveStoppingConfig = NULL,
  medianAutomatedStoppingConfig = NULL
)
}
\arguments{
\item{decayCurveStoppingConfig}{No description}

\item{medianAutomatedStoppingConfig}{No description}
}
\value{
GoogleCloudMlV1__AutomatedStoppingConfig object
}
\description{
GoogleCloudMlV1__AutomatedStoppingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for Automated Early Stopping of Trials. If no implementation_config is set, automated early stopping will not be run.
}
\concept{GoogleCloudMlV1__AutomatedStoppingConfig functions}
