% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__AutoScaling}
\alias{GoogleCloudMlV1__AutoScaling}
\title{GoogleCloudMlV1__AutoScaling Object}
\usage{
GoogleCloudMlV1__AutoScaling(maxNodes = NULL, metrics = NULL, minNodes = NULL)
}
\arguments{
\item{maxNodes}{The maximum number of nodes to scale this model under load}

\item{metrics}{MetricSpec contains the specifications to use to calculate the desired nodes count}

\item{minNodes}{Optional}
}
\value{
GoogleCloudMlV1__AutoScaling object
}
\description{
GoogleCloudMlV1__AutoScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for automatically scaling a model.
}
\concept{GoogleCloudMlV1__AutoScaling functions}
