% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__BuiltInAlgorithmOutput}
\alias{GoogleCloudMlV1__BuiltInAlgorithmOutput}
\title{GoogleCloudMlV1__BuiltInAlgorithmOutput Object}
\usage{
GoogleCloudMlV1__BuiltInAlgorithmOutput(
  framework = NULL,
  pythonVersion = NULL,
  modelPath = NULL,
  runtimeVersion = NULL
)
}
\arguments{
\item{framework}{Framework on which the built-in algorithm was trained}

\item{pythonVersion}{Python version on which the built-in algorithm was trained}

\item{modelPath}{The Cloud Storage path to the \verb{model/} directory where the training job saves the trained model}

\item{runtimeVersion}{AI Platform runtime version on which the built-in algorithm was trained}
}
\value{
GoogleCloudMlV1__BuiltInAlgorithmOutput object
}
\description{
GoogleCloudMlV1__BuiltInAlgorithmOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents output related to a built-in algorithm Job.
}
\concept{GoogleCloudMlV1__BuiltInAlgorithmOutput functions}
