% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__CancelJobRequest}
\alias{GoogleCloudMlV1__CancelJobRequest}
\title{GoogleCloudMlV1__CancelJobRequest Object}
\usage{
GoogleCloudMlV1__CancelJobRequest()
}
\value{
GoogleCloudMlV1__CancelJobRequest object
}
\description{
GoogleCloudMlV1__CancelJobRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the CancelJob method.
}
\seealso{
Other GoogleCloudMlV1__CancelJobRequest functions: 
\code{\link{projects.jobs.cancel}()}
}
\concept{GoogleCloudMlV1__CancelJobRequest functions}
