% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Capability}
\alias{GoogleCloudMlV1__Capability}
\title{GoogleCloudMlV1__Capability Object}
\usage{
GoogleCloudMlV1__Capability(type = NULL, availableAccelerators = NULL)
}
\arguments{
\item{type}{No description}

\item{availableAccelerators}{Available accelerators for the capability}
}
\value{
GoogleCloudMlV1__Capability object
}
\description{
GoogleCloudMlV1__Capability Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1__Capability functions}
