% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata}
\alias{GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata}
\title{GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata Object}
\usage{
GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata(
  createTime = NULL,
  trial = NULL,
  study = NULL
)
}
\arguments{
\item{createTime}{The time at which the operation was submitted}

\item{trial}{The trial name}

\item{study}{The name of the study that the trial belongs to}
}
\value{
GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata object
}
\description{
GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message will be placed in the metadata field of a google.longrunning.Operation associated with a CheckTrialEarlyStoppingState request.
}
\concept{GoogleCloudMlV1__CheckTrialEarlyStoppingStateMetatdata functions}
