% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse}
\alias{GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse}
\title{GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse Object}
\usage{
GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse(
  shouldStop = NULL,
  startTime = NULL,
  endTime = NULL
)
}
\arguments{
\item{shouldStop}{True if the Trial should stop}

\item{startTime}{The time at which the operation was started}

\item{endTime}{The time at which operation processing completed}
}
\value{
GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse object
}
\description{
GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The message will be placed in the response field of a completed google.longrunning.Operation associated with a CheckTrialEarlyStoppingState request.
}
\concept{GoogleCloudMlV1__CheckTrialEarlyStoppingStateResponse functions}
