% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__CompleteTrialRequest}
\alias{GoogleCloudMlV1__CompleteTrialRequest}
\title{GoogleCloudMlV1__CompleteTrialRequest Object}
\usage{
GoogleCloudMlV1__CompleteTrialRequest(
  trialInfeasible = NULL,
  infeasibleReason = NULL,
  finalMeasurement = NULL
)
}
\arguments{
\item{trialInfeasible}{Optional}

\item{infeasibleReason}{Optional}

\item{finalMeasurement}{Optional}
}
\value{
GoogleCloudMlV1__CompleteTrialRequest object
}
\description{
GoogleCloudMlV1__CompleteTrialRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the CompleteTrial service method.
}
\seealso{
Other GoogleCloudMlV1__CompleteTrialRequest functions: 
\code{\link{projects.locations.studies.trials.complete}()}
}
\concept{GoogleCloudMlV1__CompleteTrialRequest functions}
