% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Config}
\alias{GoogleCloudMlV1__Config}
\title{GoogleCloudMlV1__Config Object}
\usage{
GoogleCloudMlV1__Config(tpuServiceAccount = NULL)
}
\arguments{
\item{tpuServiceAccount}{The service account Cloud ML uses to run on TPU node}
}
\value{
GoogleCloudMlV1__Config object
}
\description{
GoogleCloudMlV1__Config Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1__Config functions}
