% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ContainerPort}
\alias{GoogleCloudMlV1__ContainerPort}
\title{GoogleCloudMlV1__ContainerPort Object}
\usage{
GoogleCloudMlV1__ContainerPort(containerPort = NULL)
}
\arguments{
\item{containerPort}{Number of the port to expose on the container}
}
\value{
GoogleCloudMlV1__ContainerPort object
}
\description{
GoogleCloudMlV1__ContainerPort Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a network port in a single container. This message is a subset of the \href{https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#containerport-v1-core}{Kubernetes ContainerPort v1 core specification}.
}
\concept{GoogleCloudMlV1__ContainerPort functions}
