% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ContainerSpec}
\alias{GoogleCloudMlV1__ContainerSpec}
\title{GoogleCloudMlV1__ContainerSpec Object}
\usage{
GoogleCloudMlV1__ContainerSpec(
  ports = NULL,
  command = NULL,
  env = NULL,
  image = NULL,
  args = NULL
)
}
\arguments{
\item{ports}{Immutable}

\item{command}{Immutable}

\item{env}{Immutable}

\item{image}{URI of the Docker image to be used as the custom container for serving predictions}

\item{args}{Immutable}
}
\value{
GoogleCloudMlV1__ContainerSpec object
}
\description{
GoogleCloudMlV1__ContainerSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specification of a custom container for serving predictions. This message is a subset of the \href{https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#container-v1-core}{Kubernetes Container v1 core specification}.
}
\concept{GoogleCloudMlV1__ContainerSpec functions}
