% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__DiskConfig}
\alias{GoogleCloudMlV1__DiskConfig}
\title{GoogleCloudMlV1__DiskConfig Object}
\usage{
GoogleCloudMlV1__DiskConfig(bootDiskType = NULL, bootDiskSizeGb = NULL)
}
\arguments{
\item{bootDiskType}{Type of the boot disk (default is 'pd-ssd')}

\item{bootDiskSizeGb}{Size in GB of the boot disk (default is 100GB)}
}
\value{
GoogleCloudMlV1__DiskConfig object
}
\description{
GoogleCloudMlV1__DiskConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the config of disk options.
}
\concept{GoogleCloudMlV1__DiskConfig functions}
