% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__EncryptionConfig}
\alias{GoogleCloudMlV1__EncryptionConfig}
\title{GoogleCloudMlV1__EncryptionConfig Object}
\usage{
GoogleCloudMlV1__EncryptionConfig(kmsKeyName = NULL)
}
\arguments{
\item{kmsKeyName}{The Cloud KMS resource identifier of the customer-managed encryption key used to protect a resource, such as a training job}
}
\value{
GoogleCloudMlV1__EncryptionConfig object
}
\description{
GoogleCloudMlV1__EncryptionConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a custom encryption key configuration that can be applied to a resource.
}
\concept{GoogleCloudMlV1__EncryptionConfig functions}
