% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__EnvVar}
\alias{GoogleCloudMlV1__EnvVar}
\title{GoogleCloudMlV1__EnvVar Object}
\usage{
GoogleCloudMlV1__EnvVar(value = NULL, name = NULL)
}
\arguments{
\item{value}{Value of the environment variable}

\item{name}{Name of the environment variable}
}
\value{
GoogleCloudMlV1__EnvVar object
}
\description{
GoogleCloudMlV1__EnvVar Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an environment variable to be made available in a container. This message is a subset of the \href{https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#envvar-v1-core}{Kubernetes EnvVar v1 core specification}.
}
\concept{GoogleCloudMlV1__EnvVar functions}
