% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ExplainRequest}
\alias{GoogleCloudMlV1__ExplainRequest}
\title{AI Platform Training & Prediction API Objects
An API to enable creating and using machine learning models.}
\usage{
GoogleCloudMlV1__ExplainRequest(httpBody = NULL)
}
\arguments{
\item{httpBody}{Required}
}
\value{
GoogleCloudMlV1__ExplainRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2022-07-13 10:40:01
filename: /Users/justin/Sync/projects/github.com/justinjm/autoGoogleAPI/googlemlv1.auto/R/ml_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GoogleCloudMlV1__ExplainRequest Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for explanations to be issued against a trained model.
}
\seealso{
Other GoogleCloudMlV1__ExplainRequest functions: 
\code{\link{projects.explain}()}
}
\concept{GoogleCloudMlV1__ExplainRequest functions}
