% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ExplanationConfig}
\alias{GoogleCloudMlV1__ExplanationConfig}
\title{GoogleCloudMlV1__ExplanationConfig Object}
\usage{
GoogleCloudMlV1__ExplanationConfig(
  sampledShapleyAttribution = NULL,
  integratedGradientsAttribution = NULL,
  xraiAttribution = NULL
)
}
\arguments{
\item{sampledShapleyAttribution}{An attribution method that approximates Shapley values for features that contribute to the label being predicted}

\item{integratedGradientsAttribution}{Attributes credit by computing the Aumann-Shapley value taking advantage of the model's fully differentiable structure}

\item{xraiAttribution}{Attributes credit by computing the XRAI taking advantage of the model's fully differentiable structure}
}
\value{
GoogleCloudMlV1__ExplanationConfig object
}
\description{
GoogleCloudMlV1__ExplanationConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message holding configuration options for explaining model predictions. There are three feature attribution methods supported for TensorFlow models: integrated gradients, sampled Shapley, and XRAI. \href{/ai-platform/prediction/docs/ai-explanations/overview}{Learn more about feature attributions.}
}
\concept{GoogleCloudMlV1__ExplanationConfig functions}
