% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__GetConfigResponse}
\alias{GoogleCloudMlV1__GetConfigResponse}
\title{GoogleCloudMlV1__GetConfigResponse Object}
\usage{
GoogleCloudMlV1__GetConfigResponse(
  serviceAccountProject = NULL,
  config = NULL,
  serviceAccount = NULL
)
}
\arguments{
\item{serviceAccountProject}{The project number for \code{service_account}}

\item{config}{No description}

\item{serviceAccount}{The service account Cloud ML uses to access resources in the project}
}
\value{
GoogleCloudMlV1__GetConfigResponse object
}
\description{
GoogleCloudMlV1__GetConfigResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Returns service account information associated with a project.
}
\concept{GoogleCloudMlV1__GetConfigResponse functions}
