% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__HyperparameterOutput}
\alias{GoogleCloudMlV1__HyperparameterOutput}
\title{GoogleCloudMlV1__HyperparameterOutput Object}
\usage{
GoogleCloudMlV1__HyperparameterOutput(
  GoogleCloudMlV1__HyperparameterOutput.hyperparameters = NULL,
  GoogleCloudMlV1__HyperparameterOutput.webAccessUris = NULL,
  builtInAlgorithmOutput = NULL,
  isTrialStoppedEarly = NULL,
  finalMetric = NULL,
  hyperparameters = NULL,
  state = NULL,
  trialId = NULL,
  webAccessUris = NULL,
  allMetrics = NULL,
  endTime = NULL,
  startTime = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__HyperparameterOutput.hyperparameters}{The \link{GoogleCloudMlV1__HyperparameterOutput.hyperparameters} object or list of objects}

\item{GoogleCloudMlV1__HyperparameterOutput.webAccessUris}{The \link{GoogleCloudMlV1__HyperparameterOutput.webAccessUris} object or list of objects}

\item{builtInAlgorithmOutput}{Details related to built-in algorithms jobs}

\item{isTrialStoppedEarly}{True if the trial is stopped early}

\item{finalMetric}{The final objective metric seen for this trial}

\item{hyperparameters}{The hyperparameters given to this trial}

\item{state}{Output only}

\item{trialId}{The trial id for these results}

\item{webAccessUris}{URIs for accessing \link{interactive shells}(https://cloud}

\item{allMetrics}{All recorded object metrics for this trial}

\item{endTime}{Output only}

\item{startTime}{Output only}
}
\value{
GoogleCloudMlV1__HyperparameterOutput object
}
\description{
GoogleCloudMlV1__HyperparameterOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the result of a single hyperparameter tuning trial from a training job. The TrainingOutput object that is returned on successful completion of a training job with hyperparameter tuning includes a list of HyperparameterOutput objects, one for each successful trial.
}
\seealso{
Other GoogleCloudMlV1__HyperparameterOutput functions: 
\code{\link{GoogleCloudMlV1__HyperparameterOutput.hyperparameters}()},
\code{\link{GoogleCloudMlV1__HyperparameterOutput.webAccessUris}()}
}
\concept{GoogleCloudMlV1__HyperparameterOutput functions}
