% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__HyperparameterOutput.webAccessUris}
\alias{GoogleCloudMlV1__HyperparameterOutput.webAccessUris}
\title{GoogleCloudMlV1__HyperparameterOutput.webAccessUris Object}
\usage{
GoogleCloudMlV1__HyperparameterOutput.webAccessUris()
}
\value{
GoogleCloudMlV1__HyperparameterOutput.webAccessUris object
}
\description{
GoogleCloudMlV1__HyperparameterOutput.webAccessUris Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
URIs for accessing \href{https://cloud.google.com/ai-platform/training/docs/monitor-debug-interactive-shell}{interactive shells} (one URI for each training node). Only available if this trial is part of a hyperparameter tuning job and the job's training_input.enable_web_access is \code{true}. The keys are names of each node in the training job; for example, \code{master-replica-0} for the master node, \code{worker-replica-0} for the first worker, and \code{ps-replica-0} for the first parameter server. The values are the URIs for each node's interactive shell.
}
\seealso{
Other GoogleCloudMlV1__HyperparameterOutput functions: 
\code{\link{GoogleCloudMlV1__HyperparameterOutput.hyperparameters}()},
\code{\link{GoogleCloudMlV1__HyperparameterOutput}()}
}
\concept{GoogleCloudMlV1__HyperparameterOutput functions}
