% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__HyperparameterSpec}
\alias{GoogleCloudMlV1__HyperparameterSpec}
\title{GoogleCloudMlV1__HyperparameterSpec Object}
\usage{
GoogleCloudMlV1__HyperparameterSpec(
  params = NULL,
  maxParallelTrials = NULL,
  algorithm = NULL,
  enableTrialEarlyStopping = NULL,
  maxFailedTrials = NULL,
  resumePreviousJobId = NULL,
  goal = NULL,
  hyperparameterMetricTag = NULL,
  maxTrials = NULL
)
}
\arguments{
\item{params}{Required}

\item{maxParallelTrials}{Optional}

\item{algorithm}{Optional}

\item{enableTrialEarlyStopping}{Optional}

\item{maxFailedTrials}{Optional}

\item{resumePreviousJobId}{Optional}

\item{goal}{Required}

\item{hyperparameterMetricTag}{Optional}

\item{maxTrials}{Optional}
}
\value{
GoogleCloudMlV1__HyperparameterSpec object
}
\description{
GoogleCloudMlV1__HyperparameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a set of hyperparameters to optimize.
}
\concept{GoogleCloudMlV1__HyperparameterSpec functions}
