% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Job.labels}
\alias{GoogleCloudMlV1__Job.labels}
\title{GoogleCloudMlV1__Job.labels Object}
\usage{
GoogleCloudMlV1__Job.labels()
}
\value{
GoogleCloudMlV1__Job.labels object
}
\description{
GoogleCloudMlV1__Job.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. One or more labels that you can add, to organize your jobs. Each label is a key-value pair, where both the key and the value are arbitrary strings that you supply. For more information, see the documentation on using labels.
}
\seealso{
Other GoogleCloudMlV1__Job functions: 
\code{\link{GoogleCloudMlV1__Job}()},
\code{\link{projects.jobs.create}()},
\code{\link{projects.jobs.patch}()}
}
\concept{GoogleCloudMlV1__Job functions}
