% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Job}
\alias{GoogleCloudMlV1__Job}
\title{GoogleCloudMlV1__Job Object}
\usage{
GoogleCloudMlV1__Job(
  GoogleCloudMlV1__Job.labels = NULL,
  state = NULL,
  startTime = NULL,
  jobId = NULL,
  trainingOutput = NULL,
  createTime = NULL,
  predictionInput = NULL,
  etag = NULL,
  labels = NULL,
  errorMessage = NULL,
  predictionOutput = NULL,
  endTime = NULL,
  trainingInput = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__Job.labels}{The \link{GoogleCloudMlV1__Job.labels} object or list of objects}

\item{state}{Output only}

\item{startTime}{Output only}

\item{jobId}{Required}

\item{trainingOutput}{The current training job result}

\item{createTime}{Output only}

\item{predictionInput}{Input parameters to create a prediction job}

\item{etag}{\code{etag} is used for optimistic concurrency control as a way to help prevent simultaneous updates of a job from overwriting each other}

\item{labels}{Optional}

\item{errorMessage}{Output only}

\item{predictionOutput}{The current prediction job result}

\item{endTime}{Output only}

\item{trainingInput}{Input parameters to create a training job}
}
\value{
GoogleCloudMlV1__Job object
}
\description{
GoogleCloudMlV1__Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a training or prediction job.
}
\seealso{
Other GoogleCloudMlV1__Job functions: 
\code{\link{GoogleCloudMlV1__Job.labels}()},
\code{\link{projects.jobs.create}()},
\code{\link{projects.jobs.patch}()}
}
\concept{GoogleCloudMlV1__Job functions}
