% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListJobsResponse}
\alias{GoogleCloudMlV1__ListJobsResponse}
\title{GoogleCloudMlV1__ListJobsResponse Object}
\usage{
GoogleCloudMlV1__ListJobsResponse(jobs = NULL, nextPageToken = NULL)
}
\arguments{
\item{jobs}{The list of jobs}

\item{nextPageToken}{Optional}
}
\value{
GoogleCloudMlV1__ListJobsResponse object
}
\description{
GoogleCloudMlV1__ListJobsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the ListJobs method.
}
\concept{GoogleCloudMlV1__ListJobsResponse functions}
