% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListLocationsResponse}
\alias{GoogleCloudMlV1__ListLocationsResponse}
\title{GoogleCloudMlV1__ListLocationsResponse Object}
\usage{
GoogleCloudMlV1__ListLocationsResponse(locations = NULL, nextPageToken = NULL)
}
\arguments{
\item{locations}{Locations where at least one type of CMLE capability is available}

\item{nextPageToken}{Optional}
}
\value{
GoogleCloudMlV1__ListLocationsResponse object
}
\description{
GoogleCloudMlV1__ListLocationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1__ListLocationsResponse functions}
