% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListModelsResponse}
\alias{GoogleCloudMlV1__ListModelsResponse}
\title{GoogleCloudMlV1__ListModelsResponse Object}
\usage{
GoogleCloudMlV1__ListModelsResponse(models = NULL, nextPageToken = NULL)
}
\arguments{
\item{models}{The list of models}

\item{nextPageToken}{Optional}
}
\value{
GoogleCloudMlV1__ListModelsResponse object
}
\description{
GoogleCloudMlV1__ListModelsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the ListModels method.
}
\concept{GoogleCloudMlV1__ListModelsResponse functions}
