% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListOptimalTrialsRequest}
\alias{GoogleCloudMlV1__ListOptimalTrialsRequest}
\title{GoogleCloudMlV1__ListOptimalTrialsRequest Object}
\usage{
GoogleCloudMlV1__ListOptimalTrialsRequest()
}
\value{
GoogleCloudMlV1__ListOptimalTrialsRequest object
}
\description{
GoogleCloudMlV1__ListOptimalTrialsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the ListTrials service method.
}
\seealso{
Other GoogleCloudMlV1__ListOptimalTrialsRequest functions: 
\code{\link{projects.locations.studies.trials.listOptimalTrials}()}
}
\concept{GoogleCloudMlV1__ListOptimalTrialsRequest functions}
