% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListOptimalTrialsResponse}
\alias{GoogleCloudMlV1__ListOptimalTrialsResponse}
\title{GoogleCloudMlV1__ListOptimalTrialsResponse Object}
\usage{
GoogleCloudMlV1__ListOptimalTrialsResponse(trials = NULL)
}
\arguments{
\item{trials}{The pareto-optimal trials for multiple objective study or the optimal trial for single objective study}
}
\value{
GoogleCloudMlV1__ListOptimalTrialsResponse object
}
\description{
GoogleCloudMlV1__ListOptimalTrialsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for the ListOptimalTrials method.
}
\concept{GoogleCloudMlV1__ListOptimalTrialsResponse functions}
