% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListTrialsResponse}
\alias{GoogleCloudMlV1__ListTrialsResponse}
\title{GoogleCloudMlV1__ListTrialsResponse Object}
\usage{
GoogleCloudMlV1__ListTrialsResponse(trials = NULL)
}
\arguments{
\item{trials}{The trials associated with the study}
}
\value{
GoogleCloudMlV1__ListTrialsResponse object
}
\description{
GoogleCloudMlV1__ListTrialsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for the ListTrials method.
}
\concept{GoogleCloudMlV1__ListTrialsResponse functions}
