% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ListVersionsResponse}
\alias{GoogleCloudMlV1__ListVersionsResponse}
\title{GoogleCloudMlV1__ListVersionsResponse Object}
\usage{
GoogleCloudMlV1__ListVersionsResponse(nextPageToken = NULL, versions = NULL)
}
\arguments{
\item{nextPageToken}{Optional}

\item{versions}{The list of versions}
}
\value{
GoogleCloudMlV1__ListVersionsResponse object
}
\description{
GoogleCloudMlV1__ListVersionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for the ListVersions method.
}
\concept{GoogleCloudMlV1__ListVersionsResponse functions}
