% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Location}
\alias{GoogleCloudMlV1__Location}
\title{GoogleCloudMlV1__Location Object}
\usage{
GoogleCloudMlV1__Location(capabilities = NULL, name = NULL)
}
\arguments{
\item{capabilities}{Capabilities available in the location}

\item{name}{No description}
}
\value{
GoogleCloudMlV1__Location object
}
\description{
GoogleCloudMlV1__Location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1__Location functions}
