% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Measurement}
\alias{GoogleCloudMlV1__Measurement}
\title{GoogleCloudMlV1__Measurement Object}
\usage{
GoogleCloudMlV1__Measurement(
  metrics = NULL,
  stepCount = NULL,
  elapsedTime = NULL
)
}
\arguments{
\item{metrics}{Provides a list of metrics that act as inputs into the objective function}

\item{stepCount}{The number of steps a machine learning model has been trained for}

\item{elapsedTime}{Output only}
}
\value{
GoogleCloudMlV1__Measurement object
}
\description{
GoogleCloudMlV1__Measurement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message representing a measurement.
}
\concept{GoogleCloudMlV1__Measurement functions}
