% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__MetricSpec}
\alias{GoogleCloudMlV1__MetricSpec}
\title{GoogleCloudMlV1__MetricSpec Object}
\usage{
GoogleCloudMlV1__MetricSpec(name = NULL, target = NULL)
}
\arguments{
\item{name}{metric name}

\item{target}{Target specifies the target value for the given metric; once real metric deviates from the threshold by a certain percentage, the node count changes}
}
\value{
GoogleCloudMlV1__MetricSpec object
}
\description{
GoogleCloudMlV1__MetricSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
MetricSpec contains the specifications to use to calculate the desired nodes count when autoscaling is enabled.
}
\concept{GoogleCloudMlV1__MetricSpec functions}
