% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Model.labels}
\alias{GoogleCloudMlV1__Model.labels}
\title{GoogleCloudMlV1__Model.labels Object}
\usage{
GoogleCloudMlV1__Model.labels()
}
\value{
GoogleCloudMlV1__Model.labels object
}
\description{
GoogleCloudMlV1__Model.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. One or more labels that you can add, to organize your models. Each label is a key-value pair, where both the key and the value are arbitrary strings that you supply. For more information, see the documentation on using labels. Note that this field is not updatable for mls1* models.
}
\seealso{
Other GoogleCloudMlV1__Model functions: 
\code{\link{GoogleCloudMlV1__Model}()},
\code{\link{projects.models.create}()},
\code{\link{projects.models.patch}()}
}
\concept{GoogleCloudMlV1__Model functions}
