% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Model}
\alias{GoogleCloudMlV1__Model}
\title{GoogleCloudMlV1__Model Object}
\usage{
GoogleCloudMlV1__Model(
  GoogleCloudMlV1__Model.labels = NULL,
  onlinePredictionConsoleLogging = NULL,
  etag = NULL,
  regions = NULL,
  labels = NULL,
  description = NULL,
  defaultVersion = NULL,
  name = NULL,
  onlinePredictionLogging = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__Model.labels}{The \link{GoogleCloudMlV1__Model.labels} object or list of objects}

\item{onlinePredictionConsoleLogging}{Optional}

\item{etag}{\code{etag} is used for optimistic concurrency control as a way to help prevent simultaneous updates of a model from overwriting each other}

\item{regions}{Optional}

\item{labels}{Optional}

\item{description}{Optional}

\item{defaultVersion}{Output only}

\item{name}{Required}

\item{onlinePredictionLogging}{Optional}
}
\value{
GoogleCloudMlV1__Model object
}
\description{
GoogleCloudMlV1__Model Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a machine learning solution. A model can have multiple versions, each of which is a deployed, trained model ready to receive prediction requests. The model itself is just a container.
}
\seealso{
Other GoogleCloudMlV1__Model functions: 
\code{\link{GoogleCloudMlV1__Model.labels}()},
\code{\link{projects.models.create}()},
\code{\link{projects.models.patch}()}
}
\concept{GoogleCloudMlV1__Model functions}
