% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__OperationMetadata.labels}
\alias{GoogleCloudMlV1__OperationMetadata.labels}
\title{GoogleCloudMlV1__OperationMetadata.labels Object}
\usage{
GoogleCloudMlV1__OperationMetadata.labels()
}
\value{
GoogleCloudMlV1__OperationMetadata.labels object
}
\description{
GoogleCloudMlV1__OperationMetadata.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user labels, inherited from the model or the model version being operated on.
}
\seealso{
Other GoogleCloudMlV1__OperationMetadata functions: 
\code{\link{GoogleCloudMlV1__OperationMetadata}()}
}
\concept{GoogleCloudMlV1__OperationMetadata functions}
