% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__OperationMetadata}
\alias{GoogleCloudMlV1__OperationMetadata}
\title{GoogleCloudMlV1__OperationMetadata Object}
\usage{
GoogleCloudMlV1__OperationMetadata(
  GoogleCloudMlV1__OperationMetadata.labels = NULL,
  projectNumber = NULL,
  operationType = NULL,
  endTime = NULL,
  isCancellationRequested = NULL,
  createTime = NULL,
  labels = NULL,
  modelName = NULL,
  version = NULL,
  startTime = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__OperationMetadata.labels}{The \link{GoogleCloudMlV1__OperationMetadata.labels} object or list of objects}

\item{projectNumber}{Contains the project number associated with the operation}

\item{operationType}{The operation type}

\item{endTime}{The time operation processing completed}

\item{isCancellationRequested}{Indicates whether a request to cancel this operation has been made}

\item{createTime}{The time the operation was submitted}

\item{labels}{The user labels, inherited from the model or the model version being operated on}

\item{modelName}{Contains the name of the model associated with the operation}

\item{version}{Contains the version associated with the operation}

\item{startTime}{The time operation processing started}
}
\value{
GoogleCloudMlV1__OperationMetadata object
}
\description{
GoogleCloudMlV1__OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the metadata of the long-running operation.
}
\seealso{
Other GoogleCloudMlV1__OperationMetadata functions: 
\code{\link{GoogleCloudMlV1__OperationMetadata.labels}()}
}
\concept{GoogleCloudMlV1__OperationMetadata functions}
