% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ParameterSpec}
\alias{GoogleCloudMlV1__ParameterSpec}
\title{GoogleCloudMlV1__ParameterSpec Object}
\usage{
GoogleCloudMlV1__ParameterSpec(
  parameterName = NULL,
  scaleType = NULL,
  discreteValues = NULL,
  maxValue = NULL,
  categoricalValues = NULL,
  type = NULL,
  minValue = NULL
)
}
\arguments{
\item{parameterName}{Required}

\item{scaleType}{Optional}

\item{discreteValues}{Required if type is \code{DISCRETE}}

\item{maxValue}{Required if type is \code{DOUBLE} or \code{INTEGER}}

\item{categoricalValues}{Required if type is \code{CATEGORICAL}}

\item{type}{Required}

\item{minValue}{Required if type is \code{DOUBLE} or \code{INTEGER}}
}
\value{
GoogleCloudMlV1__ParameterSpec object
}
\description{
GoogleCloudMlV1__ParameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single hyperparameter to optimize.
}
\concept{GoogleCloudMlV1__ParameterSpec functions}
