% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__PredictionInput}
\alias{GoogleCloudMlV1__PredictionInput}
\title{GoogleCloudMlV1__PredictionInput Object}
\usage{
GoogleCloudMlV1__PredictionInput(
  modelName = NULL,
  maxWorkerCount = NULL,
  uri = NULL,
  versionName = NULL,
  dataFormat = NULL,
  runtimeVersion = NULL,
  region = NULL,
  outputDataFormat = NULL,
  outputPath = NULL,
  batchSize = NULL,
  inputPaths = NULL,
  signatureName = NULL
)
}
\arguments{
\item{modelName}{Use this field if you want to use the default version for the specified model}

\item{maxWorkerCount}{Optional}

\item{uri}{Use this field if you want to specify a Google Cloud Storage path for the model to use}

\item{versionName}{Use this field if you want to specify a version of the model to use}

\item{dataFormat}{Required}

\item{runtimeVersion}{Optional}

\item{region}{Required}

\item{outputDataFormat}{Optional}

\item{outputPath}{Required}

\item{batchSize}{Optional}

\item{inputPaths}{Required}

\item{signatureName}{Optional}
}
\value{
GoogleCloudMlV1__PredictionInput object
}
\description{
GoogleCloudMlV1__PredictionInput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents input parameters for a prediction job.
}
\concept{GoogleCloudMlV1__PredictionInput functions}
