% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__PredictionOutput}
\alias{GoogleCloudMlV1__PredictionOutput}
\title{GoogleCloudMlV1__PredictionOutput Object}
\usage{
GoogleCloudMlV1__PredictionOutput(
  nodeHours = NULL,
  outputPath = NULL,
  predictionCount = NULL,
  errorCount = NULL
)
}
\arguments{
\item{nodeHours}{Node hours used by the batch prediction job}

\item{outputPath}{The output Google Cloud Storage location provided at the job creation time}

\item{predictionCount}{The number of generated predictions}

\item{errorCount}{The number of data instances which resulted in errors}
}
\value{
GoogleCloudMlV1__PredictionOutput object
}
\description{
GoogleCloudMlV1__PredictionOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents results of a prediction job.
}
\concept{GoogleCloudMlV1__PredictionOutput functions}
