% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__PredictRequest}
\alias{GoogleCloudMlV1__PredictRequest}
\title{GoogleCloudMlV1__PredictRequest Object}
\usage{
GoogleCloudMlV1__PredictRequest(httpBody = NULL)
}
\arguments{
\item{httpBody}{Required}
}
\value{
GoogleCloudMlV1__PredictRequest object
}
\description{
GoogleCloudMlV1__PredictRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for predictions to be issued against a trained model.
}
\seealso{
Other GoogleCloudMlV1__PredictRequest functions: 
\code{\link{projects.predict}()}
}
\concept{GoogleCloudMlV1__PredictRequest functions}
