% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__ReplicaConfig}
\alias{GoogleCloudMlV1__ReplicaConfig}
\title{GoogleCloudMlV1__ReplicaConfig Object}
\usage{
GoogleCloudMlV1__ReplicaConfig(
  imageUri = NULL,
  containerCommand = NULL,
  diskConfig = NULL,
  acceleratorConfig = NULL,
  tpuTfVersion = NULL,
  containerArgs = NULL
)
}
\arguments{
\item{imageUri}{The Docker image to run on the replica}

\item{containerCommand}{The command with which the replica's custom container is run}

\item{diskConfig}{Represents the configuration of disk options}

\item{acceleratorConfig}{Represents the type and number of accelerators used by the replica}

\item{tpuTfVersion}{The AI Platform runtime version that includes a TensorFlow version matching the one used in the custom container}

\item{containerArgs}{Arguments to the entrypoint command}
}
\value{
GoogleCloudMlV1__ReplicaConfig object
}
\description{
GoogleCloudMlV1__ReplicaConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the configuration for a replica in a cluster.
}
\concept{GoogleCloudMlV1__ReplicaConfig functions}
