% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__RequestLoggingConfig}
\alias{GoogleCloudMlV1__RequestLoggingConfig}
\title{GoogleCloudMlV1__RequestLoggingConfig Object}
\usage{
GoogleCloudMlV1__RequestLoggingConfig(
  samplingPercentage = NULL,
  bigqueryTableName = NULL
)
}
\arguments{
\item{samplingPercentage}{Percentage of requests to be logged, expressed as a fraction from 0 to 1}

\item{bigqueryTableName}{Required}
}
\value{
GoogleCloudMlV1__RequestLoggingConfig object
}
\description{
GoogleCloudMlV1__RequestLoggingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for logging request-response pairs to a BigQuery table. Online prediction requests to a model version and the responses to these requests are converted to raw strings and saved to the specified BigQuery table. Logging is constrained by \href{/bigquery/quotas}{BigQuery quotas and limits}. If your project exceeds BigQuery quotas or limits, AI Platform Prediction does not log request-response pairs, but it continues to serve predictions. If you are using \href{/ml-engine/docs/continuous-evaluation/}{continuous evaluation}, you do not need to specify this configuration manually. Setting up continuous evaluation automatically enables logging of request-response pairs.
}
\concept{GoogleCloudMlV1__RequestLoggingConfig functions}
