% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__RouteMap}
\alias{GoogleCloudMlV1__RouteMap}
\title{GoogleCloudMlV1__RouteMap Object}
\usage{
GoogleCloudMlV1__RouteMap(health = NULL, predict = NULL)
}
\arguments{
\item{health}{HTTP path on the container to send health checkss to}

\item{predict}{HTTP path on the container to send prediction requests to}
}
\value{
GoogleCloudMlV1__RouteMap object
}
\description{
GoogleCloudMlV1__RouteMap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies HTTP paths served by a custom container. AI Platform Prediction sends requests to these paths on the container; the custom container must run an HTTP server that responds to these requests with appropriate responses. Read \href{/ai-platform/prediction/docs/custom-container-requirements}{Custom container requirements} for details on how to create your container image to meet these requirements.
}
\concept{GoogleCloudMlV1__RouteMap functions}
