% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Scheduling}
\alias{GoogleCloudMlV1__Scheduling}
\title{GoogleCloudMlV1__Scheduling Object}
\usage{
GoogleCloudMlV1__Scheduling(
  maxRunningTime = NULL,
  priority = NULL,
  maxWaitTime = NULL
)
}
\arguments{
\item{maxRunningTime}{Optional}

\item{priority}{Optional}

\item{maxWaitTime}{Optional}
}
\value{
GoogleCloudMlV1__Scheduling object
}
\description{
GoogleCloudMlV1__Scheduling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
All parameters related to scheduling of training jobs.
}
\concept{GoogleCloudMlV1__Scheduling functions}
