% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__SetDefaultVersionRequest}
\alias{GoogleCloudMlV1__SetDefaultVersionRequest}
\title{GoogleCloudMlV1__SetDefaultVersionRequest Object}
\usage{
GoogleCloudMlV1__SetDefaultVersionRequest()
}
\value{
GoogleCloudMlV1__SetDefaultVersionRequest object
}
\description{
GoogleCloudMlV1__SetDefaultVersionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for the SetDefaultVersion request.
}
\seealso{
Other GoogleCloudMlV1__SetDefaultVersionRequest functions: 
\code{\link{projects.models.versions.setDefault}()}
}
\concept{GoogleCloudMlV1__SetDefaultVersionRequest functions}
