% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__StopTrialRequest}
\alias{GoogleCloudMlV1__StopTrialRequest}
\title{GoogleCloudMlV1__StopTrialRequest Object}
\usage{
GoogleCloudMlV1__StopTrialRequest()
}
\value{
GoogleCloudMlV1__StopTrialRequest object
}
\description{
GoogleCloudMlV1__StopTrialRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GoogleCloudMlV1__StopTrialRequest functions: 
\code{\link{projects.locations.studies.trials.stop}()}
}
\concept{GoogleCloudMlV1__StopTrialRequest functions}
