% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Study}
\alias{GoogleCloudMlV1__Study}
\title{GoogleCloudMlV1__Study Object}
\usage{
GoogleCloudMlV1__Study(studyConfig = NULL)
}
\arguments{
\item{studyConfig}{Required}
}
\value{
GoogleCloudMlV1__Study object
}
\description{
GoogleCloudMlV1__Study Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message representing a Study.
}
\seealso{
Other GoogleCloudMlV1__Study functions: 
\code{\link{projects.locations.studies.create}()}
}
\concept{GoogleCloudMlV1__Study functions}
