% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__StudyConfig}
\alias{GoogleCloudMlV1__StudyConfig}
\title{GoogleCloudMlV1__StudyConfig Object}
\usage{
GoogleCloudMlV1__StudyConfig(
  parameters = NULL,
  algorithm = NULL,
  automatedStoppingConfig = NULL,
  metrics = NULL
)
}
\arguments{
\item{parameters}{Required}

\item{algorithm}{The search algorithm specified for the study}

\item{automatedStoppingConfig}{Configuration for automated stopping of unpromising Trials}

\item{metrics}{Metric specs for the study}
}
\value{
GoogleCloudMlV1__StudyConfig object
}
\description{
GoogleCloudMlV1__StudyConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents configuration of a study.
}
\concept{GoogleCloudMlV1__StudyConfig functions}
