% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__SuggestTrialsMetadata}
\alias{GoogleCloudMlV1__SuggestTrialsMetadata}
\title{GoogleCloudMlV1__SuggestTrialsMetadata Object}
\usage{
GoogleCloudMlV1__SuggestTrialsMetadata(
  createTime = NULL,
  study = NULL,
  suggestionCount = NULL,
  clientId = NULL
)
}
\arguments{
\item{createTime}{The time operation was submitted}

\item{study}{The name of the study that the trial belongs to}

\item{suggestionCount}{The number of suggestions requested}

\item{clientId}{The identifier of the client that is requesting the suggestion}
}
\value{
GoogleCloudMlV1__SuggestTrialsMetadata object
}
\description{
GoogleCloudMlV1__SuggestTrialsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata field of a google.longrunning.Operation associated with a SuggestTrialsRequest.
}
\concept{GoogleCloudMlV1__SuggestTrialsMetadata functions}
