% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__SuggestTrialsRequest}
\alias{GoogleCloudMlV1__SuggestTrialsRequest}
\title{GoogleCloudMlV1__SuggestTrialsRequest Object}
\usage{
GoogleCloudMlV1__SuggestTrialsRequest(suggestionCount = NULL, clientId = NULL)
}
\arguments{
\item{suggestionCount}{Required}

\item{clientId}{Required}
}
\value{
GoogleCloudMlV1__SuggestTrialsRequest object
}
\description{
GoogleCloudMlV1__SuggestTrialsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the SuggestTrial service method.
}
\seealso{
Other GoogleCloudMlV1__SuggestTrialsRequest functions: 
\code{\link{projects.locations.studies.trials.suggest}()}
}
\concept{GoogleCloudMlV1__SuggestTrialsRequest functions}
