% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__SuggestTrialsResponse}
\alias{GoogleCloudMlV1__SuggestTrialsResponse}
\title{GoogleCloudMlV1__SuggestTrialsResponse Object}
\usage{
GoogleCloudMlV1__SuggestTrialsResponse(
  studyState = NULL,
  trials = NULL,
  endTime = NULL,
  startTime = NULL
)
}
\arguments{
\item{studyState}{The state of the study}

\item{trials}{A list of trials}

\item{endTime}{The time at which operation processing completed}

\item{startTime}{The time at which the operation was started}
}
\value{
GoogleCloudMlV1__SuggestTrialsResponse object
}
\description{
GoogleCloudMlV1__SuggestTrialsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message will be placed in the response field of a completed google.longrunning.Operation associated with a SuggestTrials request.
}
\concept{GoogleCloudMlV1__SuggestTrialsResponse functions}
