% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__TrainingInput}
\alias{GoogleCloudMlV1__TrainingInput}
\title{GoogleCloudMlV1__TrainingInput Object}
\usage{
GoogleCloudMlV1__TrainingInput(
  encryptionConfig = NULL,
  runtimeVersion = NULL,
  jobDir = NULL,
  args = NULL,
  enableWebAccess = NULL,
  masterConfig = NULL,
  hyperparameters = NULL,
  masterType = NULL,
  scheduling = NULL,
  packageUris = NULL,
  workerCount = NULL,
  serviceAccount = NULL,
  workerType = NULL,
  parameterServerCount = NULL,
  evaluatorCount = NULL,
  workerConfig = NULL,
  pythonVersion = NULL,
  scaleTier = NULL,
  evaluatorType = NULL,
  region = NULL,
  parameterServerType = NULL,
  evaluatorConfig = NULL,
  pythonModule = NULL,
  network = NULL,
  useChiefInTfConfig = NULL,
  parameterServerConfig = NULL
)
}
\arguments{
\item{encryptionConfig}{Optional}

\item{runtimeVersion}{Optional}

\item{jobDir}{Optional}

\item{args}{Optional}

\item{enableWebAccess}{Optional}

\item{masterConfig}{Optional}

\item{hyperparameters}{Optional}

\item{masterType}{Optional}

\item{scheduling}{Optional}

\item{packageUris}{Required}

\item{workerCount}{Optional}

\item{serviceAccount}{Optional}

\item{workerType}{Optional}

\item{parameterServerCount}{Optional}

\item{evaluatorCount}{Optional}

\item{workerConfig}{Optional}

\item{pythonVersion}{Optional}

\item{scaleTier}{Required}

\item{evaluatorType}{Optional}

\item{region}{Required}

\item{parameterServerType}{Optional}

\item{evaluatorConfig}{Optional}

\item{pythonModule}{Required}

\item{network}{Optional}

\item{useChiefInTfConfig}{Optional}

\item{parameterServerConfig}{Optional}
}
\value{
GoogleCloudMlV1__TrainingInput object
}
\description{
GoogleCloudMlV1__TrainingInput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents input parameters for a training job. When using the gcloud command to submit your training job, you can specify the input parameters as command-line arguments and/or in a YAML configuration file referenced from the --config command-line argument. For details, see the guide to \href{/ai-platform/training/docs/training-jobs}{submitting a training job}.
}
\concept{GoogleCloudMlV1__TrainingInput functions}
