% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__TrainingOutput}
\alias{GoogleCloudMlV1__TrainingOutput}
\title{GoogleCloudMlV1__TrainingOutput Object}
\usage{
GoogleCloudMlV1__TrainingOutput(
  GoogleCloudMlV1__TrainingOutput.webAccessUris = NULL,
  trials = NULL,
  isHyperparameterTuningJob = NULL,
  consumedMLUnits = NULL,
  hyperparameterMetricTag = NULL,
  isBuiltInAlgorithmJob = NULL,
  builtInAlgorithmOutput = NULL,
  completedTrialCount = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__TrainingOutput.webAccessUris}{The \link{GoogleCloudMlV1__TrainingOutput.webAccessUris} object or list of objects}

\item{trials}{Results for individual Hyperparameter trials}

\item{isHyperparameterTuningJob}{Whether this job is a hyperparameter tuning job}

\item{consumedMLUnits}{The amount of ML units consumed by the job}

\item{hyperparameterMetricTag}{The TensorFlow summary tag name used for optimizing hyperparameter tuning trials}

\item{isBuiltInAlgorithmJob}{Whether this job is a built-in Algorithm job}

\item{builtInAlgorithmOutput}{Details related to built-in algorithms jobs}

\item{completedTrialCount}{The number of hyperparameter tuning trials that completed successfully}
}
\value{
GoogleCloudMlV1__TrainingOutput object
}
\description{
GoogleCloudMlV1__TrainingOutput Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents results of a training job. Output only.
}
\seealso{
Other GoogleCloudMlV1__TrainingOutput functions: 
\code{\link{GoogleCloudMlV1__TrainingOutput.webAccessUris}()}
}
\concept{GoogleCloudMlV1__TrainingOutput functions}
