% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__TrainingOutput.webAccessUris}
\alias{GoogleCloudMlV1__TrainingOutput.webAccessUris}
\title{GoogleCloudMlV1__TrainingOutput.webAccessUris Object}
\usage{
GoogleCloudMlV1__TrainingOutput.webAccessUris()
}
\value{
GoogleCloudMlV1__TrainingOutput.webAccessUris object
}
\description{
GoogleCloudMlV1__TrainingOutput.webAccessUris Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Output only. URIs for accessing \href{https://cloud.google.com/ai-platform/training/docs/monitor-debug-interactive-shell}{interactive shells} (one URI for each training node). Only available if training_input.enable_web_access is \code{true}. The keys are names of each node in the training job; for example, \code{master-replica-0} for the master node, \code{worker-replica-0} for the first worker, and \code{ps-replica-0} for the first parameter server. The values are the URIs for each node's interactive shell.
}
\seealso{
Other GoogleCloudMlV1__TrainingOutput functions: 
\code{\link{GoogleCloudMlV1__TrainingOutput}()}
}
\concept{GoogleCloudMlV1__TrainingOutput functions}
