% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Trial}
\alias{GoogleCloudMlV1__Trial}
\title{GoogleCloudMlV1__Trial Object}
\usage{
GoogleCloudMlV1__Trial(
  finalMeasurement = NULL,
  state = NULL,
  measurements = NULL,
  parameters = NULL
)
}
\arguments{
\item{finalMeasurement}{The final measurement containing the objective value}

\item{state}{The detailed state of a trial}

\item{measurements}{A list of measurements that are strictly lexicographically ordered by their induced tuples (steps, elapsed_time)}

\item{parameters}{The parameters of the trial}
}
\value{
GoogleCloudMlV1__Trial object
}
\description{
GoogleCloudMlV1__Trial Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message representing a trial.
}
\seealso{
Other GoogleCloudMlV1__Trial functions: 
\code{\link{projects.locations.studies.trials.create}()}
}
\concept{GoogleCloudMlV1__Trial functions}
