% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__Version}
\alias{GoogleCloudMlV1__Version}
\title{GoogleCloudMlV1__Version Object}
\usage{
GoogleCloudMlV1__Version(
  GoogleCloudMlV1__Version.labels = NULL,
  acceleratorConfig = NULL,
  errorMessage = NULL,
  manualScaling = NULL,
  labels = NULL,
  requestLoggingConfig = NULL,
  serviceAccount = NULL,
  machineType = NULL,
  isDefault = NULL,
  container = NULL,
  predictionClass = NULL,
  etag = NULL,
  name = NULL,
  createTime = NULL,
  lastUseTime = NULL,
  runtimeVersion = NULL,
  deploymentUri = NULL,
  state = NULL,
  explanationConfig = NULL,
  autoScaling = NULL,
  framework = NULL,
  packageUris = NULL,
  description = NULL,
  routes = NULL,
  pythonVersion = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__Version.labels}{The \link{GoogleCloudMlV1__Version.labels} object or list of objects}

\item{acceleratorConfig}{Optional}

\item{errorMessage}{Output only}

\item{manualScaling}{Manually select the number of nodes to use for serving the model}

\item{labels}{Optional}

\item{requestLoggingConfig}{Optional}

\item{serviceAccount}{Optional}

\item{machineType}{Optional}

\item{isDefault}{Output only}

\item{container}{Optional}

\item{predictionClass}{Optional}

\item{etag}{\code{etag} is used for optimistic concurrency control as a way to help prevent simultaneous updates of a model from overwriting each other}

\item{name}{Required}

\item{createTime}{Output only}

\item{lastUseTime}{Output only}

\item{runtimeVersion}{Required}

\item{deploymentUri}{The Cloud Storage URI of a directory containing trained model artifacts to be used to create the model version}

\item{state}{Output only}

\item{explanationConfig}{Optional}

\item{autoScaling}{Automatically scale the number of nodes used to serve the model in response to increases and decreases in traffic}

\item{framework}{Optional}

\item{packageUris}{Optional}

\item{description}{Optional}

\item{routes}{Optional}

\item{pythonVersion}{Required}
}
\value{
GoogleCloudMlV1__Version object
}
\description{
GoogleCloudMlV1__Version Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a version of the model. Each version is a trained model deployed in the cloud, ready to handle prediction requests. A model can have multiple versions. You can get information about all of the versions of a given model by calling projects.models.versions.list.
}
\seealso{
Other GoogleCloudMlV1__Version functions: 
\code{\link{GoogleCloudMlV1__Version.labels}()},
\code{\link{projects.models.versions.create}()},
\code{\link{projects.models.versions.patch}()}
}
\concept{GoogleCloudMlV1__Version functions}
