% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1__XraiAttribution}
\alias{GoogleCloudMlV1__XraiAttribution}
\title{GoogleCloudMlV1__XraiAttribution Object}
\usage{
GoogleCloudMlV1__XraiAttribution(numIntegralSteps = NULL)
}
\arguments{
\item{numIntegralSteps}{Number of steps for approximating the path integral}
}
\value{
GoogleCloudMlV1__XraiAttribution object
}
\description{
GoogleCloudMlV1__XraiAttribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Attributes credit by computing the XRAI taking advantage of the model's fully differentiable structure. Refer to this paper for more details: https://arxiv.org/abs/1906.02825 Currently only implemented for models with natural image inputs.
}
\concept{GoogleCloudMlV1__XraiAttribution functions}
