% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig}
\alias{GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig}
\title{GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig Object}
\usage{
GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig(
  useElapsedTime = NULL
)
}
\arguments{
\item{useElapsedTime}{If true, measurement}
}
\value{
GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig object
}
\description{
GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1_AutomatedStoppingConfig_DecayCurveAutomatedStoppingConfig functions}
