% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig}
\alias{GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig}
\title{GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig Object}
\usage{
GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig(
  useElapsedTime = NULL
)
}
\arguments{
\item{useElapsedTime}{If true, the median automated stopping rule applies to measurement}
}
\value{
GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig object
}
\description{
GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The median automated stopping rule stops a pending trial if the trial's best objective_value is strictly below the median 'performance' of all completed trials reported up to the trial's last measurement. Currently, 'performance' refers to the running average of the objective values reported by the trial in each measurement.
}
\concept{GoogleCloudMlV1_AutomatedStoppingConfig_MedianAutomatedStoppingConfig functions}
