% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_Measurement_Metric}
\alias{GoogleCloudMlV1_Measurement_Metric}
\title{GoogleCloudMlV1_Measurement_Metric Object}
\usage{
GoogleCloudMlV1_Measurement_Metric(value = NULL, metric = NULL)
}
\arguments{
\item{value}{Required}

\item{metric}{Required}
}
\value{
GoogleCloudMlV1_Measurement_Metric object
}
\description{
GoogleCloudMlV1_Measurement_Metric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message representing a metric in the measurement.
}
\concept{GoogleCloudMlV1_Measurement_Metric functions}
