% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfig_MetricSpec}
\alias{GoogleCloudMlV1_StudyConfig_MetricSpec}
\title{GoogleCloudMlV1_StudyConfig_MetricSpec Object}
\usage{
GoogleCloudMlV1_StudyConfig_MetricSpec(goal = NULL, metric = NULL)
}
\arguments{
\item{goal}{Required}

\item{metric}{Required}
}
\value{
GoogleCloudMlV1_StudyConfig_MetricSpec object
}
\description{
GoogleCloudMlV1_StudyConfig_MetricSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a metric to optimize.
}
\concept{GoogleCloudMlV1_StudyConfig_MetricSpec functions}
