% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfig_ParameterSpec}
\alias{GoogleCloudMlV1_StudyConfig_ParameterSpec}
\title{GoogleCloudMlV1_StudyConfig_ParameterSpec Object}
\usage{
GoogleCloudMlV1_StudyConfig_ParameterSpec(
  discreteValueSpec = NULL,
  type = NULL,
  scaleType = NULL,
  parameter = NULL,
  childParameterSpecs = NULL,
  doubleValueSpec = NULL,
  integerValueSpec = NULL,
  parentIntValues = NULL,
  categoricalValueSpec = NULL,
  parentDiscreteValues = NULL,
  parentCategoricalValues = NULL
)
}
\arguments{
\item{discreteValueSpec}{The value spec for a 'DISCRETE' parameter}

\item{type}{Required}

\item{scaleType}{How the parameter should be scaled}

\item{parameter}{Required}

\item{childParameterSpecs}{A child node is active if the parameter's value matches the child node's matching_parent_values}

\item{doubleValueSpec}{The value spec for a 'DOUBLE' parameter}

\item{integerValueSpec}{The value spec for an 'INTEGER' parameter}

\item{parentIntValues}{No description}

\item{categoricalValueSpec}{The value spec for a 'CATEGORICAL' parameter}

\item{parentDiscreteValues}{No description}

\item{parentCategoricalValues}{No description}
}
\value{
GoogleCloudMlV1_StudyConfig_ParameterSpec object
}
\description{
GoogleCloudMlV1_StudyConfig_ParameterSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single parameter to optimize.
}
\concept{GoogleCloudMlV1_StudyConfig_ParameterSpec functions}
