% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec(values = NULL)
}
\arguments{
\item{values}{Must be specified if type is \code{CATEGORICAL}}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_CategoricalValueSpec functions}
