% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec(values = NULL)
}
\arguments{
\item{values}{Must be specified if type is \code{DISCRETE}}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_DiscreteValueSpec functions}
