% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec(
  minValue = NULL,
  maxValue = NULL
)
}
\arguments{
\item{minValue}{Must be specified if type is \code{DOUBLE}}

\item{maxValue}{Must be specified if type is \code{DOUBLE}}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_DoubleValueSpec functions}
