% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec(
  minValue = NULL,
  maxValue = NULL
)
}
\arguments{
\item{minValue}{Must be specified if type is \code{INTEGER}}

\item{maxValue}{Must be specified if type is \code{INTEGER}}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_IntegerValueSpec functions}
