% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec(
  values = NULL
)
}
\arguments{
\item{values}{Matches values of the parent parameter with type 'CATEGORICAL'}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec to match categorical values from parent parameter.
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentCategoricalValueSpec functions}
