% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec(
  values = NULL
)
}
\arguments{
\item{values}{Matches values of the parent parameter with type 'DISCRETE'}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec to match discrete values from parent parameter.
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentDiscreteValueSpec functions}
