% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec}
\alias{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec}
\title{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec Object}
\usage{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec(
  values = NULL
)
}
\arguments{
\item{values}{Matches values of the parent parameter with type 'INTEGER'}
}
\value{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec object
}
\description{
GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the spec to match integer values from parent parameter.
}
\concept{GoogleCloudMlV1_StudyConfigParameterSpec_MatchingParentIntValueSpec functions}
