% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleIamV1__Binding}
\alias{GoogleIamV1__Binding}
\title{GoogleIamV1__Binding Object}
\usage{
GoogleIamV1__Binding(members = NULL, role = NULL, condition = NULL)
}
\arguments{
\item{members}{Specifies the principals requesting access for a Google Cloud resource}

\item{role}{Role that is assigned to the list of \code{members}, or principals}

\item{condition}{The condition that is associated with this binding}
}
\value{
GoogleIamV1__Binding object
}
\description{
GoogleIamV1__Binding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Associates \code{members}, or principals, with a \code{role}.
}
\concept{GoogleIamV1__Binding functions}
