% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleIamV1__SetIamPolicyRequest}
\alias{GoogleIamV1__SetIamPolicyRequest}
\title{GoogleIamV1__SetIamPolicyRequest Object}
\usage{
GoogleIamV1__SetIamPolicyRequest(policy = NULL, updateMask = NULL)
}
\arguments{
\item{policy}{REQUIRED: The complete policy to be applied to the \code{resource}}

\item{updateMask}{OPTIONAL: A FieldMask specifying which fields of the policy to modify}
}
\value{
GoogleIamV1__SetIamPolicyRequest object
}
\description{
GoogleIamV1__SetIamPolicyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for \code{SetIamPolicy} method.
}
\seealso{
Other GoogleIamV1__SetIamPolicyRequest functions: 
\code{\link{projects.jobs.setIamPolicy}()},
\code{\link{projects.models.setIamPolicy}()}
}
\concept{GoogleIamV1__SetIamPolicyRequest functions}
