% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleIamV1__TestIamPermissionsRequest}
\alias{GoogleIamV1__TestIamPermissionsRequest}
\title{GoogleIamV1__TestIamPermissionsRequest Object}
\usage{
GoogleIamV1__TestIamPermissionsRequest(permissions = NULL)
}
\arguments{
\item{permissions}{The set of permissions to check for the \code{resource}}
}
\value{
GoogleIamV1__TestIamPermissionsRequest object
}
\description{
GoogleIamV1__TestIamPermissionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for \code{TestIamPermissions} method.
}
\seealso{
Other GoogleIamV1__TestIamPermissionsRequest functions: 
\code{\link{projects.jobs.testIamPermissions}()},
\code{\link{projects.models.testIamPermissions}()}
}
\concept{GoogleIamV1__TestIamPermissionsRequest functions}
