% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleLongrunning__ListOperationsResponse}
\alias{GoogleLongrunning__ListOperationsResponse}
\title{GoogleLongrunning__ListOperationsResponse Object}
\usage{
GoogleLongrunning__ListOperationsResponse(
  nextPageToken = NULL,
  operations = NULL
)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{operations}{A list of operations that matches the specified filter in the request}
}
\value{
GoogleLongrunning__ListOperationsResponse object
}
\description{
GoogleLongrunning__ListOperationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Operations.ListOperations.
}
\concept{GoogleLongrunning__ListOperationsResponse functions}
