% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleLongrunning__Operation}
\alias{GoogleLongrunning__Operation}
\title{GoogleLongrunning__Operation Object}
\usage{
GoogleLongrunning__Operation(
  GoogleLongrunning__Operation.response = NULL,
  GoogleLongrunning__Operation.metadata = NULL,
  done = NULL,
  error = NULL,
  name = NULL,
  response = NULL,
  metadata = NULL
)
}
\arguments{
\item{GoogleLongrunning__Operation.response}{The \link{GoogleLongrunning__Operation.response} object or list of objects}

\item{GoogleLongrunning__Operation.metadata}{The \link{GoogleLongrunning__Operation.metadata} object or list of objects}

\item{done}{If the value is \code{false}, it means the operation is still in progress}

\item{error}{The error result of the operation in case of failure or cancellation}

\item{name}{The server-assigned name, which is only unique within the same service that originally returns it}

\item{response}{The normal response of the operation in case of success}

\item{metadata}{Service-specific metadata associated with the operation}
}
\value{
GoogleLongrunning__Operation object
}
\description{
GoogleLongrunning__Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This resource represents a long-running operation that is the result of a network API call.
}
\seealso{
Other GoogleLongrunning__Operation functions: 
\code{\link{GoogleLongrunning__Operation.metadata}()},
\code{\link{GoogleLongrunning__Operation.response}()}
}
\concept{GoogleLongrunning__Operation functions}
