% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleLongrunning__Operation.response}
\alias{GoogleLongrunning__Operation.response}
\title{GoogleLongrunning__Operation.response Object}
\usage{
GoogleLongrunning__Operation.response()
}
\value{
GoogleLongrunning__Operation.response object
}
\description{
GoogleLongrunning__Operation.response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The normal response of the operation in case of success. If the original method returns no data on success, such as \code{Delete}, the response is \code{google.protobuf.Empty}. If the original method is standard \code{Get}/\code{Create}/\code{Update}, the response should be the resource. For other methods, the response should have the type \code{XxxResponse}, where \code{Xxx} is the original method name. For example, if the original method name is \code{TakeSnapshot()}, the inferred response type is \code{TakeSnapshotResponse}.
}
\seealso{
Other GoogleLongrunning__Operation functions: 
\code{\link{GoogleLongrunning__Operation.metadata}()},
\code{\link{GoogleLongrunning__Operation}()}
}
\concept{GoogleLongrunning__Operation functions}
