% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_objects.R
\name{GoogleRpc__Status}
\alias{GoogleRpc__Status}
\title{GoogleRpc__Status Object}
\usage{
GoogleRpc__Status(code = NULL, details = NULL, message = NULL)
}
\arguments{
\item{code}{The status code, which should be an enum value of google}

\item{details}{A list of messages that carry the error details}

\item{message}{A developer-facing error message, which should be in English}
}
\value{
GoogleRpc__Status object
}
\description{
GoogleRpc__Status Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The \code{Status} type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by \href{https://github.com/grpc}{gRPC}. Each \code{Status} message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the \href{https://cloud.google.com/apis/design/errors}{API Design Guide}.
}
\concept{GoogleRpc__Status functions}
