% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\docType{package}
\name{ml_googleAuthR}
\alias{ml_googleAuthR}
\title{AI Platform Training & Prediction API
An API to enable creating and using machine learning models.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
at 2022-07-13 10:40:00
filename: /Users/justin/Sync/projects/github.com/justinjm/autoGoogleAPI/googlemlv1.auto/R/ml_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}
}
