% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.explain}
\alias{projects.explain}
\title{Performs explanation on the data in the request. {\% dynamic include '/ai-platform/includes/___explain-request' \%}}
\usage{
projects.explain(GoogleCloudMlV1__ExplainRequest, name)
}
\arguments{
\item{GoogleCloudMlV1__ExplainRequest}{The \link{GoogleCloudMlV1__ExplainRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__ExplainRequest functions: 
\code{\link{GoogleCloudMlV1__ExplainRequest}()}
}
\concept{GoogleCloudMlV1__ExplainRequest functions}
