% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.jobs.cancel}
\alias{projects.jobs.cancel}
\title{Cancels a running job.}
\usage{
projects.jobs.cancel(GoogleCloudMlV1__CancelJobRequest, name)
}
\arguments{
\item{GoogleCloudMlV1__CancelJobRequest}{The \link{GoogleCloudMlV1__CancelJobRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__CancelJobRequest functions: 
\code{\link{GoogleCloudMlV1__CancelJobRequest}()}
}
\concept{GoogleCloudMlV1__CancelJobRequest functions}
