% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.jobs.create}
\alias{projects.jobs.create}
\title{Creates a training or a batch prediction job.}
\usage{
projects.jobs.create(GoogleCloudMlV1__Job, parent)
}
\arguments{
\item{GoogleCloudMlV1__Job}{The \link{GoogleCloudMlV1__Job} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__Job functions: 
\code{\link{GoogleCloudMlV1__Job.labels}()},
\code{\link{GoogleCloudMlV1__Job}()},
\code{\link{projects.jobs.patch}()}
}
\concept{GoogleCloudMlV1__Job functions}
