% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.jobs.list}
\alias{projects.jobs.list}
\title{Lists the jobs in the project. If there are no jobs that match the request parameters, the list request returns an empty response body: {}.}
\usage{
projects.jobs.list(parent, pageSize = NULL, pageToken = NULL, filter = NULL)
}
\arguments{
\item{parent}{Required}

\item{pageSize}{Optional}

\item{pageToken}{Optional}

\item{filter}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}
}
