% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.locations.studies.trials.listOptimalTrials}
\alias{projects.locations.studies.trials.listOptimalTrials}
\title{Lists the pareto-optimal trials for multi-objective study or the optimal trials for single-objective study. The definition of pareto-optimal can be checked in wiki page. https://en.wikipedia.org/wiki/Pareto_efficiency}
\usage{
projects.locations.studies.trials.listOptimalTrials(
  GoogleCloudMlV1__ListOptimalTrialsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudMlV1__ListOptimalTrialsRequest}{The \link{GoogleCloudMlV1__ListOptimalTrialsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__ListOptimalTrialsRequest functions: 
\code{\link{GoogleCloudMlV1__ListOptimalTrialsRequest}()}
}
\concept{GoogleCloudMlV1__ListOptimalTrialsRequest functions}
