% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.locations.studies.trials.stop}
\alias{projects.locations.studies.trials.stop}
\title{Stops a trial.}
\usage{
projects.locations.studies.trials.stop(GoogleCloudMlV1__StopTrialRequest, name)
}
\arguments{
\item{GoogleCloudMlV1__StopTrialRequest}{The \link{GoogleCloudMlV1__StopTrialRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__StopTrialRequest functions: 
\code{\link{GoogleCloudMlV1__StopTrialRequest}()}
}
\concept{GoogleCloudMlV1__StopTrialRequest functions}
