% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.locations.studies.trials.suggest}
\alias{projects.locations.studies.trials.suggest}
\title{Adds one or more trials to a study, with parameter values suggested by AI Platform Vizier. Returns a long-running operation associated with the generation of trial suggestions. When this long-running operation succeeds, it will contain a SuggestTrialsResponse.}
\usage{
projects.locations.studies.trials.suggest(
  GoogleCloudMlV1__SuggestTrialsRequest,
  parent
)
}
\arguments{
\item{GoogleCloudMlV1__SuggestTrialsRequest}{The \link{GoogleCloudMlV1__SuggestTrialsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__SuggestTrialsRequest functions: 
\code{\link{GoogleCloudMlV1__SuggestTrialsRequest}()}
}
\concept{GoogleCloudMlV1__SuggestTrialsRequest functions}
