% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.create}
\alias{projects.models.create}
\title{Creates a model which will later contain one or more versions. You must add at least one version before you can request predictions from the model. Add versions by calling projects.models.versions.create.}
\usage{
projects.models.create(GoogleCloudMlV1__Model, parent)
}
\arguments{
\item{GoogleCloudMlV1__Model}{The \link{GoogleCloudMlV1__Model} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__Model functions: 
\code{\link{GoogleCloudMlV1__Model.labels}()},
\code{\link{GoogleCloudMlV1__Model}()},
\code{\link{projects.models.patch}()}
}
\concept{GoogleCloudMlV1__Model functions}
