% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.getIamPolicy}
\alias{projects.models.getIamPolicy}
\title{Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.}
\usage{
projects.models.getIamPolicy(resource, options.requestedPolicyVersion = NULL)
}
\arguments{
\item{resource}{REQUIRED: The resource for which the policy is being requested}

\item{options.requestedPolicyVersion}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}
}
