% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.patch}
\alias{projects.models.patch}
\title{Updates a specific model resource. Currently the only supported fields to update are \code{description} and \code{default_version.name}.}
\usage{
projects.models.patch(GoogleCloudMlV1__Model, name, updateMask = NULL)
}
\arguments{
\item{GoogleCloudMlV1__Model}{The \link{GoogleCloudMlV1__Model} object to pass to this method}

\item{name}{Required}

\item{updateMask}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__Model functions: 
\code{\link{GoogleCloudMlV1__Model.labels}()},
\code{\link{GoogleCloudMlV1__Model}()},
\code{\link{projects.models.create}()}
}
\concept{GoogleCloudMlV1__Model functions}
