% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.setIamPolicy}
\alias{projects.models.setIamPolicy}
\title{Sets the access control policy on the specified resource. Replaces any existing policy. Can return \code{NOT_FOUND}, \code{INVALID_ARGUMENT}, and \code{PERMISSION_DENIED} errors.}
\usage{
projects.models.setIamPolicy(GoogleIamV1__SetIamPolicyRequest, resource)
}
\arguments{
\item{GoogleIamV1__SetIamPolicyRequest}{The \link{GoogleIamV1__SetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being specified}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleIamV1__SetIamPolicyRequest functions: 
\code{\link{GoogleIamV1__SetIamPolicyRequest}()},
\code{\link{projects.jobs.setIamPolicy}()}
}
\concept{GoogleIamV1__SetIamPolicyRequest functions}
