% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.versions.create}
\alias{projects.models.versions.create}
\title{Creates a new version of a model from a trained TensorFlow model. If the version created in the cloud by this call is the first deployed version of the specified model, it will be made the default version of the model. When you add a version to a model that already has one or more versions, the default version does not automatically change. If you want a new version to be the default, you must call projects.models.versions.setDefault.}
\usage{
projects.models.versions.create(GoogleCloudMlV1__Version, parent)
}
\arguments{
\item{GoogleCloudMlV1__Version}{The \link{GoogleCloudMlV1__Version} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__Version functions: 
\code{\link{GoogleCloudMlV1__Version.labels}()},
\code{\link{GoogleCloudMlV1__Version}()},
\code{\link{projects.models.versions.patch}()}
}
\concept{GoogleCloudMlV1__Version functions}
