% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.versions.patch}
\alias{projects.models.versions.patch}
\title{Updates the specified Version resource. Currently the only update-able fields are \code{description}, \code{requestLoggingConfig}, \code{autoScaling.minNodes}, and \code{manualScaling.nodes}.}
\usage{
projects.models.versions.patch(
  GoogleCloudMlV1__Version,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleCloudMlV1__Version}{The \link{GoogleCloudMlV1__Version} object to pass to this method}

\item{name}{Required}

\item{updateMask}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__Version functions: 
\code{\link{GoogleCloudMlV1__Version.labels}()},
\code{\link{GoogleCloudMlV1__Version}()},
\code{\link{projects.models.versions.create}()}
}
\concept{GoogleCloudMlV1__Version functions}
