% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{projects.models.versions.setDefault}
\alias{projects.models.versions.setDefault}
\title{Designates a version to be the default for the model. The default version is used for prediction requests made against the model that don't specify a version. The first version to be created for a model is automatically set as the default. You must make any subsequent changes to the default version setting manually using this method.}
\usage{
projects.models.versions.setDefault(
  GoogleCloudMlV1__SetDefaultVersionRequest,
  name
)
}
\arguments{
\item{GoogleCloudMlV1__SetDefaultVersionRequest}{The \link{GoogleCloudMlV1__SetDefaultVersionRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/ml/}{Google Documentation}

Other GoogleCloudMlV1__SetDefaultVersionRequest functions: 
\code{\link{GoogleCloudMlV1__SetDefaultVersionRequest}()}
}
\concept{GoogleCloudMlV1__SetDefaultVersionRequest functions}
